#28: resolve bundled xsl/dictionaries by walking ancestors, not a fixed hop count
Both bundled-resource lookups hardcoded three ".." hops from the assembly directory, which silently
assumed the RID-less output layout bin/<Config>/<tfm>. Any RID-specific build adds a segment
(bin/<Config>/<tfm>/<rid>), so three hops landed on bin/ and the directory was never found. The app
then fell through to the empty per-user directory it had just created, and every book rendered
"XSL file not found".

Replace the fixed hop with a bounded upward search for the named directory, shared by both call
sites via BundledResourceLocator so the subtle part exists in one place. Precedence is unchanged:
dev source tree, then beside the executable, then the macOS .app Resources dir.

Also fixes the XSL branch searching for "Xsl" - the real directory is lowercase "xsl", and the
mismatch only ever worked because Windows and default APFS are case-insensitive.
